Fixed mali failing to report EPROBE_DEFER
In the module initialization for job manager, it should return
error if any of the initialization it calls return error.

The static variable is persistent even if the module is
unloaded. After module unload, sscd_dev contains garbage.
Made a constant copy of sscd_dev and sscd_dev be initialized
every time the module is loaded.

Bug: 218395503
Test: test patch simulating defer situation
Change-Id: I26070c6b2c8e826b2fc78fe337cdf0ca1d2f154b